+++ Packages/BrandonUtils/Standalone/Strings/PrettificationSettings.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 
 using BrandonUtils.Standalone.Optional;
+using BrandonUtils.Standalone.Strings.Json;
 
 using JetBrains.Annotations;
