+++ Packages/BrandonUtils/Tests/Standalone/Strings/PrettificationTests.cs
@@ -9,6 +9,7 @@ using BrandonUtils.Standalone.Collections;
 using BrandonUtils.Standalone.Enums;
 using BrandonUtils.Standalone.Reflection;
 using BrandonUtils.Standalone.Strings;
+using BrandonUtils.Standalone.Strings.Json;
 using BrandonUtils.Standalone.Strings.Prettifiers;
 using BrandonUtils.Testing;
 using BrandonUtils.Tests.Standalone.Reflection;
